refactor(bot): split monolithic bot.py into focused modules by responsibility
- Extract keyboard builders into bot_keyboards.py with dedicated tests
- Move command handlers to bot_commands.py with comprehensive test coverage
- Separate callback logic into bot_callbacks.py with proper import handling
- Reduce bot.py to core entry point functions and application wiring
- Update all imports between modules and remove unused dependencies
- Relocate tests to co-locate with their target modules for better organization
- Maintain all existing functionality while improving code structure
