+++ src/api/bot.py
@@ -29,17 +29,15 @@ from telegram.ext import (
     filters,
 )
 
+from src.api.bot_callbacks import handle_callback
+from src.api.bot_commands import (
+    help_command,
+    reminders_command,
+    start_command,
+    timezone_command,
+)
 from src.api.bot_helpers import truncate_for_telegram
 from src.core.config import get_settings
-from src.db.user_settings import (
-    add_reminder,
-    get_reminders,
-    get_user_settings_id,
-    insert_default_reminders,
-    remove_reminder,
-    update_timezone,
-    upsert_user_settings,
-)
 
 logger = logging.getLogger(__name__)
 
@@ -63,227 +61,6 @@ def _get_graph(context: ContextTypes.DEFAULT_TYPE) -> CompiledStateGraph:
             checkpointer=context.bot_data["saver"],
         )
     return context.bot_data["graph"]
-
-
-def _build_timezone_keyboard(chat_id: int) -> InlineKeyboardMarkup:
-    """Build an inline keyboard with UTC offset buttons.
-
-    Args:
-        chat_id: The Telegram chat ID to encode in callback data.
-
-    Returns:
-        InlineKeyboardMarkup with UTC offset buttons.
-    """
-    rows: list[list[InlineKeyboardButton]] = []
-    offsets = list(range(-12, 15))  # UTC-12 through UTC+14
-    row: list[InlineKeyboardButton] = []
-    for offset in offsets:
-        label = f"UTC{offset:+d}" if offset != 0 else "UTC+0"
-        row.append(InlineKeyboardButton(label, callback_data=f"tz|{offset}|{chat_id}"))
-        if len(row) == 4:
-            rows.append(row)
-            row = []
-    if row:
-        rows.append(row)
-    return InlineKeyboardMarkup(rows)
-
-
-def _build_reminders_message(
-    reminders: list[dict[str, str]],
-    user_settings_id: str,
-) -> tuple[str, InlineKeyboardMarkup]:
-    """Build the reminders display text and inline keyboard.
-
-    Args:
-        reminders: List of reminder dicts with id and time.
-        user_settings_id: The user_settings UUID for callback data.
-
-    Returns:
-        Tuple of (message text, inline keyboard markup).
-    """
-    if not reminders:
-        text = "⏰ You have no reminders set."
-    else:
-        lines = ["⏰ Your reminders:"]
-        for r in reminders:
-            time_display = r["time"][:5]
-            lines.append(f"  • {time_display}")
-        text = "\n".join(lines)
-
-    buttons: list[list[InlineKeyboardButton]] = []
-    for idx, r in enumerate(reminders):
-        time_display = r["time"][:5]
-        buttons.append(
-            [
-                InlineKeyboardButton(
-                    f"❌ Remove {time_display}",
-                    callback_data=f"rm_rem|{idx}",
-                )
-            ]
-        )
-    if len(reminders) < 5:
-        buttons.append(
-            [
-                InlineKeyboardButton(
-                    "➕ Add reminder",
-                    callback_data="add_rem",
-                )
-            ]
-        )
-
-    return text, InlineKeyboardMarkup(buttons)
-
-
-def _build_hour_picker_keyboard(user_settings_id: str) -> InlineKeyboardMarkup:
-    """Build an inline keyboard with hour buttons (00:00 - 23:00).
-
-    Args:
-        user_settings_id: The user_settings UUID for callback data.
-
-    Returns:
-        InlineKeyboardMarkup with hour buttons.
-    """
-    rows: list[list[InlineKeyboardButton]] = []
-    row: list[InlineKeyboardButton] = []
-    for hour in range(24):
-        label = f"{hour:02d}:00"
-        row.append(InlineKeyboardButton(label, callback_data=f"add_hr|{hour}"))
-        if len(row) == 4:
-            rows.append(row)
-            row = []
-    if row:
-        rows.append(row)
-    return InlineKeyboardMarkup(rows)
-
-
-async def start_command(
-    update: Update,
-    context: ContextTypes.DEFAULT_TYPE,
-) -> None:
-    """Handle the /start command.
-
-    Upserts user settings and shows a timezone picker for new users.
-
-    Args:
-        update: The Telegram update.
-        context: The Telegram context.
-    """
-    chat_id = str(update.message.chat.id)  # type: ignore[union-attr]
-    is_new = upsert_user_settings(chat_id)
-
-    await update.message.reply_text(  # type: ignore[union-attr]
-        """🧠 <b>Welcome to ThinkBack</b>
-
-Your space to capture ideas, revisit them, and actually remember what matters.
-
-<b>Here's how you can use this space:</b>
-• 💡 Share an insight → I'll help you save it
-• 🔍 Ask questions → I'll search your saved knowledge
-• 💬 Just chat → I'm here for that too
-
-<b>⏰ Spaced Repetition</b>
-Saved insights come back to you later (default is 12:00 noon) with reflections to strengthen your memory.
-You can customize reminder times anytime with /reminders.
-
-<b>Quick setup:</b>
-• 🌍 Set your timezone: /timezone
-• ❓ See all commands: /help
-
-Start by sharing your first thought 👇
-""",  # noqa: E501
-        parse_mode=ParseMode.HTML,
-    )
-
-    if is_new:
-        user_settings_id = get_user_settings_id(chat_id)
-        if user_settings_id:
-            insert_default_reminders(user_settings_id)
-
-        keyboard = _build_timezone_keyboard(update.message.chat.id)  # type: ignore[union-attr]
-        await update.message.reply_text(  # type: ignore[union-attr]
-            "🌍 What's your UTC offset?",
-            reply_markup=keyboard,
-        )
-
-
-async def timezone_command(
-    update: Update,
-    context: ContextTypes.DEFAULT_TYPE,
-) -> None:
-    """Handle the /timezone command.
-
-    Shows the UTC offset picker so the user can update their timezone.
-
-    Args:
-        update: The Telegram update.
-        context: The Telegram context.
-    """
-    chat_id = update.message.chat.id  # type: ignore[union-attr]
-    keyboard = _build_timezone_keyboard(chat_id)
-    await update.message.reply_text(  # type: ignore[union-attr]
-        "🌍 Select your UTC offset:",
-        reply_markup=keyboard,
-    )
-
-
-async def reminders_command(
-    update: Update,
-    context: ContextTypes.DEFAULT_TYPE,
-) -> None:
-    """Handle the /reminders command.
-
-    Shows current reminder times with options to add or remove.
-
-    Args:
-        update: The Telegram update.
-        context: The Telegram context.
-    """
-    chat_id = str(update.message.chat.id)  # type: ignore[union-attr]
-    user_settings_id = get_user_settings_id(chat_id)
-
-    if not user_settings_id:
-        await update.message.reply_text(  # type: ignore[union-attr]
-            "Please run /start first to set up your account."
-        )
-        return
-
-    reminders = get_reminders(user_settings_id)
-    text, keyboard = _build_reminders_message(reminders, user_settings_id)
-    await update.message.reply_text(  # type: ignore[union-attr]
-        text,
-        reply_markup=keyboard,
-    )
-
-
-async def help_command(
-    update: Update,
-    context: ContextTypes.DEFAULT_TYPE,
-) -> None:
-    """Handle the /help command.
-
-    Lists all available bot commands.
-
-    Args:
-        update: The Telegram update.
-        context: The Telegram context.
-    """
-    await update.message.reply_text(  # type: ignore[union-attr]
-        """📚 <b>Help & Commands</b>
-
-<b>Core commands</b>
-/start — Get started or reset your intro
-/timezone — Set your local timezone
-/reminders — Customize when insights resurface
-/help — Show this guide
-
-<b>How to use ThinkBack</b>
-• 💡 Share an insight → I'll help you save it
-• 🔍 Ask a question → I'll answer using your saved insights
-
-What do you want to try first?
-""",
-        parse_mode=ParseMode.HTML,
-    )
 
 
 async def handle_message(
@@ -427,137 +204,6 @@ async def handle_message(
                     safe_response,
                     parse_mode=ParseMode.HTML,
                 )
-
-
-async def handle_callback(
-    update: Update,
-    context: ContextTypes.DEFAULT_TYPE,
-) -> None:
-    """Handle inline button callbacks for save confirmation.
-
-    Args:
-        update: The Telegram update containing the callback query.
-        context: The Telegram context.
-    """
-    query = update.callback_query
-    if not query or not query.data:
-        return
-
-    await query.answer()
-
-    parts = query.data.split("|")
-    action = parts[0]
-
-    if action == "tz":
-        # Timezone selection: callback_data = "tz|<offset>|<chat_id>"
-        offset = int(parts[1])
-        chat_id = parts[2]
-        # POSIX convention: Etc/GMT-7 means UTC+7 (inverted sign)
-        if offset == 0:
-            tz_str = "UTC"
-        else:
-            tz_str = f"Etc/GMT{-offset:+d}"
-        update_timezone(chat_id, tz_str)
-        label = f"UTC{offset:+d}" if offset != 0 else "UTC+0"
-        try:
-            await query.edit_message_text(text=f"Timezone set to {label} ✅")
-        except TelegramError:
-            logger.exception("Failed to edit timezone message")
-        return
-
-    if action == "rm_rem":
-        # Remove reminder: callback_data = "rm_rem|<index>"
-        if not query.message:
-            return
-        chat_id = str(query.message.chat.id)
-        user_settings_id = get_user_settings_id(chat_id)
-        if not user_settings_id:
-            return
-        reminders = get_reminders(user_settings_id)
-        idx = int(parts[1])
-        if idx < len(reminders):
-            remove_reminder(reminders[idx]["id"])
-        reminders = get_reminders(user_settings_id)
-        text, keyboard = _build_reminders_message(reminders, user_settings_id)
-        try:
-            await query.edit_message_text(text=text, reply_markup=keyboard)
-        except TelegramError:
-            logger.exception("Failed to edit reminders message")
-        return
-
-    if action == "add_rem":
-        # Show hour picker: callback_data = "add_rem"
-        if not query.message:
-            return
-        chat_id = str(query.message.chat.id)
-        user_settings_id = get_user_settings_id(chat_id)
-        if not user_settings_id:
-            return
-        keyboard = _build_hour_picker_keyboard(user_settings_id)
-        try:
-            await query.edit_message_text(
-                text="🕐 Select a time for the new reminder:",
-                reply_markup=keyboard,
-            )
-        except TelegramError:
-            logger.exception("Failed to edit hour picker message")
-        return
-
-    if action == "add_hr":
-        # Add reminder at selected hour: callback_data = "add_hr|<hour>"
-        if not query.message:
-            return
-        chat_id = str(query.message.chat.id)
-        user_settings_id = get_user_settings_id(chat_id)
-        if not user_settings_id:
-            return
-        hour = int(parts[1])
-        time_str = f"{hour:02d}:00"
-        added = add_reminder(user_settings_id, time_str)
-        if not added:
-            try:
-                await query.edit_message_text(text="⚠️ Maximum 5 reminders reached.")
-            except TelegramError:
-                logger.exception("Failed to edit max reminders message")
-            return
-        reminders = get_reminders(user_settings_id)
-        text, keyboard = _build_reminders_message(reminders, user_settings_id)
-        try:
-            await query.edit_message_text(text=text, reply_markup=keyboard)
-        except TelegramError:
-            logger.exception("Failed to edit reminders message")
-        return
-
-    # Save confirmation: callback_data = "save_yes|<thread_id>" or "save_no|<thread_id>"
-    thread_id = parts[1] if len(parts) > 1 else ""
-    approved = action == "save_yes"
-
-    graph = _get_graph(context)
-    config = RunnableConfig({"configurable": {"thread_id": thread_id}})
-
-    from langgraph.types import Command
-
-    try:
-        result = await graph.ainvoke(
-            Command(resume={"approved": approved}),
-            config=config,
-        )
-
-        messages = result.get("messages", [])
-        response = messages[-1].content if messages and hasattr(messages[-1], "content") else ""
-        response = response or ("Memory saved! ✅" if approved else "Save cancelled. ❌")
-
-    except Exception:
-        logger.exception("Error processing callback")
-        response = "Sorry, something went wrong."
-
-    try:
-        await query.edit_message_text(
-            text=truncate_for_telegram(response),
-            parse_mode=ParseMode.HTML,
-        )
-    except TelegramError:
-        logger.exception("Failed to edit callback message")
 
 
 def create_application() -> Application:
